requirements.txt
torch
torchvision
gradio
pillow
tqdm
scikit-learn
